requirements.txt
pyperclip
selenium
python-dotenv